% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choose_blank_prev}
\alias{choose_blank_prev}
\title{Choose the “previous” blank for a given sample folder}
\usage{
choose_blank_prev(sample_fname, blank_folders)
}
\arguments{
\item{sample_fname}{Basename of the sample folder (e.g. "059-…_TESTPEPTIDEAK.D")}

\item{blank_folders}{Character vector of full paths to blank .D folders}
}
\value{
Single path (or NULL)
}
\description{
Choose the “previous” blank for a given sample folder
}
