% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_top_peaks}
\alias{filter_top_peaks}
\title{Filter and rank peaks}
\usage{
filter_top_peaks(peaks_tbl, min_rt = 6, top_n = 10)
}
\arguments{
\item{peaks_tbl}{Tibble from detect_peaks_on_smoothed}

\item{min_rt}{Numeric. Minimum apex_rt to keep.}

\item{top_n}{Integer. How many highest peaks to return.}
}
\value{
A filtered & arranged tibble
}
\description{
Filter and rank peaks
}
