% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agilent.R
\name{read_hplc_agilent}
\alias{read_hplc_agilent}
\title{Read Agilent `.D` folder (raw trace)}
\usage{
read_hplc_agilent(d_path, signal_wavelength = 214, signal_ref = 360)
}
\arguments{
\item{d_path}{Character. Path to the Agilent `.D` directory.}

\item{signal_wavelength}{Numeric. DAD signal wavelength (nm). Default 214.}

\item{signal_ref}{Numeric. DAD reference wavelength (nm). Default 360.}
}
\value{
A data.frame with columns:
  - **time** (minutes)
  - **intensity** (mAU)
}
\description{
Read Agilent `.D` folder (raw trace)
}
\seealso{
\code{\link{baseline_als}}, \code{\link{baseline_hybrid_sm}}
}
