% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_epsilon_214.R
\name{estimate_epsilon_214}
\alias{estimate_epsilon_214}
\title{Estimate the sequence‐specific ε214 (Kuipers & Gruppen, 2007),
with special tripeptide corrections}
\usage{
estimate_epsilon_214(sequence)
}
\arguments{
\item{sequence}{Character. Peptide one‐letter code.}
}
\value{
Numeric. ε214 in M⁻¹·cm⁻¹.
}
\description{
Estimate the sequence‐specific ε214 (Kuipers & Gruppen, 2007),
with special tripeptide corrections
}
